finalizado

+++ src/model/Audio.java
@@ -4,7 +4,7 @@ public class Audio {
     private String title;
     private int totalReproductions;
     private int totalLikes;
-    private double classification;
+    private int classification;
 
     public String getTitle() {
         return title;
@@ -22,7 +22,7 @@ public class Audio {
         return totalLikes;
     }
 
-    public double getClassification() {
+    public int getClassification() {
         return classification;
     }
